# Improve mince and add method for non-uniform partition of IntervalBox