Closes #444