@lbenet 2021-05-29
Some benchmarking:

- Current master:
```
julia> @benchmark mince(ii, ndiv) setup=(ii=-1 .. 1, ndiv=100_000_000)
BenchmarkTools.Trial: 
  memory estimate:  152.59 MiB
  allocs estimate:  2
  --------------
  minimum time:     45.498 ms (0.00% GC)
  median time:      64.020 ms (26.33% GC)
  mean time:        58.819 ms (17.82% GC)
  maximum time:     120.155 ms (58.98% GC)
  --------------
  samples:          85
  evals/sample:     1
```

- This PR
```
julia> @benchmark mince(ii, ndiv) setup=(ii=-1 .. 1, ndiv=100_000_000)
BenchmarkTools.Trial: 
  memory estimate:  152.59 MiB
  allocs estimate:  2
  --------------
  minimum time:     41.535 ms (0.00% GC)
  median time:      53.344 ms (16.84% GC)
  mean time:        54.540 ms (19.06% GC)
  maximum time:     117.183 ms (60.92% GC)
  --------------
  samples:          92
  evals/sample:     1
```